% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.GreyZone.R
\name{runExample.GreyZone}
\alias{runExample.GreyZone}
\title{Run example of the Grey Zone method, simulating a large variety
of tests with varying qualities.}
\usage{
runExample.GreyZone()
}
\description{
Documentation is provided in the ShinyApp itself.
}
\seealso{
\code{\link{runExample.UI}} \code{\link{runExample.TGROC}}
\code{\link{runExample.ROC}}
}
