% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.ROC.R
\name{runExample.ROC}
\alias{runExample.ROC}
\title{Run example of the ROC method, simulating a large variety
of tests with varying qualities.}
\usage{
runExample.ROC()
}
\description{
Documentation is provided in the ShinyApp itself.
}
\seealso{
\code{\link{runExample.UI}} \code{\link{runExample.TGROC}}
\code{\link{runExample.GreyZone}}
}
