% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.TGROC.R
\name{runExample.TGROC}
\alias{runExample.TGROC}
\title{Run example of the TG-ROC method, simulating a large variety
of tests with varying qualities.}
\usage{
runExample.TGROC()
}
\description{
Documentation is provided in the ShinyApp itself.
}
\seealso{
\code{\link{runExample.UI}} \code{\link{runExample.GreyZone}}
\code{\link{runExample.ROC}}
}
