% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.UI.R
\name{runExample.UI}
\alias{runExample.UI}
\title{Run example of the UI (UncertainInterval method), simulating a large variety
of tests with varying qualities.}
\usage{
runExample.UI()
}
\description{
Documentation is provided in the ShinyApp itself.
}
\seealso{
\code{\link{runExample.TGROC}} \code{\link{runExample.GreyZone}}
\code{\link{runExample.ROC}}
}
